Rewrote bucket sort to use modern C++
The current version of the bucket sort uses C style arrays and index
based for loops.
This commit modifies the bucket sort algorithm to use modern C++
features such as vectors and range based for loops as well as auto for
type specifiers.

Furthermore, the `using namespace std;` line has been removed as it is
generally considered a bad practice.